fix(bench): 细化基准指标展示精度并记录真机测试反馈
Benchmark 面板区分单次/末轮与平均 TPS，展示 prefill/decode 拆分；
native 毫秒指标改为小数避免 TTFT 显示为 0；新增真机模拟器测试反馈文档。

+++ android/app/src/main/cpp/eyes_llm_session.cpp
@@ -317,9 +317,10 @@ std::string EyesLlmSession::infer(
 
     const auto* context = llm->getContext();
     if (context != nullptr) {
-        metrics_["ttft_ms"] = std::to_string(context->ttfa_us / 1000);
-        metrics_["prefill_ms"] = std::to_string(context->prefill_us / 1000);
-        metrics_["decode_ms"] = std::to_string(context->decode_us / 1000);
+        // 保留小数 ms，避免 ttfa_us<1000 被整除成 0 导致 UI 显示 TTFT=0
+        metrics_["ttft_ms"] = std::to_string(static_cast<double>(context->ttfa_us) / 1000.0);
+        metrics_["prefill_ms"] = std::to_string(static_cast<double>(context->prefill_us) / 1000.0);
+        metrics_["decode_ms"] = std::to_string(static_cast<double>(context->decode_us) / 1000.0);
         metrics_["vision_ms"] = std::to_string(context->vision_us / 1000);
         metrics_["audio_ms"] = std::to_string(context->audio_us / 1000);
         metrics_["tokens_generated"] = std::to_string(context->gen_seq_len);
